Move Log and TailOptions to separate file
When Log and TailOptions are used by other classes it is now more
obvious that they're not hard lined to Tail.
